Moved py wrappers from main.py to db_operations.py
